requirements.txt
onnxruntime
numpy
Pillow
fastapi
mangum
uvicorn
python-jose
python-dotenv
scikit-learn
opencv-python
cmake
dlib==19.24.2
scikit-image
requests